+++ .storybook/theme.js
@@ -1,5 +1,6 @@
 import { create } from "@storybook/theming";
 import LOGO from "./nimbus-logo.png";
+import tokens from "@nimbus-ds/tokens/dist/js/tokens";
 
 export const base = {
   base: "light",
@@ -11,4 +12,15 @@ export const base = {
     "CentraNube, -apple-system, BlinkMacSystemFont, Roboto, 'Helvetica Neue', arial, sans-serif",
 };
 
-export default create(base);
+export const darkExtend = {
+  base: "dark",
+
+  appBg: tokens.color.sys.dark.neutral.background.value,
+  appContentBg: tokens.color.sys.dark.neutral.surface.value,
+
+  barBg: tokens.color.sys.dark.neutral.surface.value,
+
+  textColor: tokens.color.sys.dark.neutral["text-high"].value,
+};
+
+export default create(base, darkExtend);